Add HAProxy integration.
This makes it possible to deploy HAProxy on e.g. the master node, allowing
clients and applications to use a primary and possibly a standby port for
read-write or read-only queries.

The inventory needs a "[haproxy_server]" group, and the PostgreSQL, Patroni
RAFT and HAProxy write/read ports need to be defined.

The variable haproxy_primary_read_scale decides whether the primary should take
part in queries sent to the read-only port or not.